prontuarioeletronico/services/auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pytest==8.2.0
httpx==0.27.0
PyJWT==2.10.1
